@@ -33,11 +33,12 @@
     "postcss-flexbugs-fixes": "3.2.0",
     "postcss-loader": "2.0.6",
     "promise": "8.0.1",
-    "react": "^15.6.1",
+    "react": "^16.7.0",
     "react-dev-utils": "^4.0.1",
-    "react-dom": "^15.6.1",
+    "react-dom": "^16.7.0",
+    "react-loadable": "^5.5.0",
     "react-router-dom": "^4.2.2",
-    "semantic-ui-css": "^2.2.12",
+    "semantic-ui-css": "2.2.14",
     "style-loader": "0.18.2",
     "sw-precache-webpack-plugin": "0.11.4",
     "url-loader": "0.5.9",
