@@ -45,7 +45,7 @@ class AddMedicine extends Component {
           data: newLog
         });
 
-        this.props.history.push("/tabs/medicines");
+        this.props.history.push("/app/medicines");
       });
   }
 
@@ -102,7 +102,7 @@ class AddMedicine extends Component {
             </Modal.Description>
           </Modal.Content>
           <Modal.Actions>
-            <Link to="/tabs/medicines" className="ui button negative">
+            <Link to="/app/medicines" className="ui button negative">
               Cancel
             </Link>
             <Button onClick={this.onAdd} positive>
