@@ -94,7 +94,7 @@ class AddChart extends Component {
           data: newLog
         });
 
-        this.props.history.push("/tabs/patients/");
+        this.props.history.push("/app/patients/");
       });
   }
 
@@ -157,7 +157,7 @@ class AddChart extends Component {
             </Form>
           </Modal.Content>
           <Modal.Actions>
-            <Link to="/tabs/patients/" className="ui button negative">
+            <Link to="/app/patients/" className="ui button negative">
               Cancel
             </Link>
             <Button onClick={this.onAdd} positive>
