@@ -56,7 +56,7 @@ class AddPatient extends Component {
           data: newLog
         });
 
-        this.props.history.push("/tabs/patients");
+        this.props.history.push("/app/patients");
       });
   }
 
@@ -172,7 +172,7 @@ class AddPatient extends Component {
             </Form>
           </Modal.Content>
           <Modal.Actions>
-            <Link to="/tabs/patients/" className="ui button negative">
+            <Link to="/app/patients/" className="ui button negative">
               Cancel
             </Link>
             <Button onClick={this.onAdd} positive>
