@@ -47,7 +47,7 @@ class AddPrescription extends Component {
           url: "http://localhost:3001/api/logs/",
           data: newLog
         });
-        this.props.history.push("/tabs/patients/");
+        this.props.history.push("/app/patients/");
       });
   }
 
@@ -97,7 +97,7 @@ class AddPrescription extends Component {
             </Modal.Description>
           </Modal.Content>
           <Modal.Actions>
-            <Link to="/tabs/patients" className="ui button negative">
+            <Link to="/app/patients" className="ui button negative">
               Cancel
             </Link>
             <Button onClick={this.onAdd} positive>
