@@ -45,7 +45,7 @@ class DeleteImage extends Component {
           data: newLog
         });
 
-        this.props.history.push("/tabs/patients/view/" + this.state.patientId);
+        this.props.history.push("/app/patients/view/" + this.state.patientId);
       });
   }
 
@@ -81,7 +81,7 @@ class DeleteImage extends Component {
           </Modal.Content>
           <Modal.Actions>
             <Link
-              to={"/tabs/patients/view/" + this.state.patientId}
+              to={"/app/patients/view/" + this.state.patientId}
               className="ui button negative"
             >
               Cancel
