@@ -44,7 +44,7 @@ class DeletePatient extends Component {
           data: newLog
         });
 
-        this.props.history.push("/tabs/patients");
+        this.props.history.push("/app/patients");
       });
   }
 
@@ -71,7 +71,7 @@ class DeletePatient extends Component {
             </Modal.Description>
           </Modal.Content>
           <Modal.Actions>
-            <Link to="/tabs/patients" className="ui button negative">
+            <Link to="/app/patients" className="ui button negative">
               Cancel
             </Link>
             <Button onClick={this.onDelete} positive>
