@@ -71,7 +71,7 @@ class EditPatient extends Component {
           data: newLog
         });
 
-        this.props.history.push("/tabs/patients");
+        this.props.history.push("/app/patients");
       });
   }
 
@@ -207,7 +207,7 @@ class EditPatient extends Component {
             </Form>
           </Modal.Content>
           <Modal.Actions>
-            <Link to="/tabs/patients/" className="ui button negative">
+            <Link to="/app/patients/" className="ui button negative">
               Cancel
             </Link>
             <Button onClick={this.onEdit} positive>
