@@ -45,7 +45,7 @@ class AddProcedure extends Component {
           data: newLog
         });
 
-        this.props.history.push("/tabs/procedures");
+        this.props.history.push("/app/procedures");
       });
   }
 
@@ -105,7 +105,7 @@ class AddProcedure extends Component {
             </Modal.Description>
           </Modal.Content>
           <Modal.Actions>
-            <Link to="/tabs/procedures/" className="ui button negative">
+            <Link to="/app/procedures/" className="ui button negative">
               Cancel
             </Link>
             <Button onClick={this.onAdd} positive>
